Rename mispelt mete -> meta file
